+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Changelog
 Unreleased
 ~~~~~~~~~~
 
+- Introduce ``ANY_STATE`` and ``ANY_OTHER_STATE`` constants
+
 django-fsm-2 4.2.3 2026-03-15
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~